requirements.txt
click==8.1.3
colorama==0.4.6
markdown-it-py==2.2.0
mdurl==0.1.2
Pillow==9.5.0
pydantic==1.10.7
pyee==9.1.0
Pygments==2.15.1
python-ffmpeg==2.0.4
rich==13.3.5
shellingham==1.5.0.post1
typer==0.9.0
typing_extensions==4.5.0
